Batches 150-153: QE mined schema, BoltzTraP2 transport, latent-defect sweep, tracked headless harness
Batch 150: Mine QE 7.2-7.6 INPUT_PW/PH/HP definitions into generated
QESchemaData (572 entries, per-literal version masks); QENamelistSchema,
QESchemaValidator, QESchemaAudit, ResultAnalysisService mined-audit
section, miner script, QE_INTEGRATION_ROADMAP doc, executed tests.

Batch 152: BoltzTraP2TraceParser ground-truth rewrite (kind by data-row
count, halltens refusal, Fortran 0/4/8 tensor offsets, verbatim units),
ResultAnalysisService BOLTZTRAP2_TRANSPORT analysis kind end-to-end,
11 executed tests.

Batch 153: Fix 12 latent master defects proven by real javac + executed
tests: QEBrillouinZoneGeometry compile error + face scrambling,
SymmetricEigen3 rotation type error, ProjectProperty blank-context ctor,
ResultAnalysisService unclosed paren, QEMdDiffusionMsdParser missing
import, CubeGridReader multi-catch, WorkflowAudit undeclared field,
QEDynmatModesParser decoration lines, Wannier90 spread tolerance flake,
plus 9 test-file repairs; new SymmetricEigen3ResidualTest (5 tests).

Infra: tracked headless JDK17 harness at scripts/harness/ (mini-JUnit5
stubs, MiniRunner, QE structural stubs, setup/run scripts) - full suite
now 150 classes / 530 tests / 0 failures executed headlessly; pins in
compile_check.py; renamed comparison doc brand-clean.

+++ tests/java/quantumforge/run/parser/TrajectoryIndexReaderTest.java
@@ -40,10 +40,10 @@ class TrajectoryIndexReaderTest {
         assertFalse(index.isTruncatedTail());
         assertTrue(index.isOffsetsComplete());
         assertEquals(3, index.getOffsets().size());
-        assertEquals(0L, index.getOffsets().get(0));
+        assertEquals(0L, index.getOffsets().get(0).longValue());
         int frameBytes = FRAME.getBytes(StandardCharsets.UTF_8).length;
-        assertEquals((long) frameBytes, index.getOffsets().get(1));
-        assertEquals(2L * frameBytes, index.getOffsets().get(2));
+        assertEquals((long) frameBytes, index.getOffsets().get(1).longValue());
+        assertEquals(2L * frameBytes, index.getOffsets().get(2).longValue());
     }
 
     @Test
